fix :fix#14 fare calculation based on distance
